SPINOZA-1751 [geohash-java] Extend GeoHash to have a (relatively short) string representation that is valid for all bit-precision levels (not just those evenly divisible by 5)
Modified the encoding and decoding schemes, and added
a new unit test for the GeoHash class.

This encoding right-pads the GeoHash with 0s when its precision
is not already a multiple of 5 bits.

To differentiate between GeoHash strings that have different
bits precision yet map to the same value, use the
"GeoHash.fromGeohashString(String geohash, int numberOfBits)"
method.

reviewed by @drmathochist